Squelch level detector (first version) working
Squelch level detector (first version) working. For transmitter to be
active the input signal must be the double of noise level and be present
during at least 2 seconds. Transmitter will be active for 2 minutes. It
is automatically renewed if input signal still meets activation
conditions.
ADC reference level set to 2.048 instead of 1.024V.